Added separated test helper contracts for gov and ibc.
