+++ src/test_helpers/mod.rs
@@ -7,10 +7,11 @@ use serde::{Deserialize, Serialize};
 pub mod caller;
 pub mod echo;
 pub mod error;
+pub mod gov;
 pub mod hackatom;
+pub mod ibc;
 pub mod payout;
 pub mod reflect;
-pub mod stargate;
 
 /// Custom message for testing purposes.
 #[derive(Debug, Clone, Serialize, Deserialize, JsonSchema, PartialEq, Eq)]
